requirements.txt
# Core dependencies
ollama>=0.1.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Speech interface
faster-whisper>=0.10.0
piper-tts>=1.2.0
silero-vad>=4.0.0
pyaudio>=0.2.13
sounddevice>=0.4.6
pvporcupine>=3.0.0

# LLM & embeddings
sentence-transformers>=2.2.0
chromadb>=0.4.0

# Database & storage
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Web dashboard
flask>=3.0.0
flask-cors>=4.0.0
websockets>=12.0

# Utilities
requests>=2.31.0
aiohttp>=3.9.0
numpy>=1.24.0
pandas>=2.0.0
rich>=13.0.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
